requirements.txt
streamlit==1.43.1
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
statsmodels==0.14.0
matplotlib==3.7.2
seaborn==0.13.2

jinja2==3.1.2
python-docx==0.8.11
requests==2.31.0
python-dotenv==1.0.1
openai==0.27.8
ollama==0.4.7

pycaret==3.0.4
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.2
scikit-plot==0.3.7
imbalanced-learn==0.12.4
plotly==5.24.1
